requirements.txt
# 网络诊断工具依赖包

# GUI框架（Python标准库，无需安装）
# tkinter - 已包含在Python标准安装中

# 系统和网络工具（Python标准库，无需安装）
# socket, subprocess, platform, threading - 标准库

# 性能监控和系统信息
psutil>=5.9.0

# PDF报告生成
reportlab>=3.6.0

# 网络拓扑可视化和热力图
matplotlib>=3.5.0
networkx>=2.6.0
scipy>=1.9.0
pillow>=9.0.0

# 数据处理和分析 (实时监控优化)
pandas>=1.3.0
numpy>=1.21.0

# 配置文件管理（企业级功能）
pyyaml>=6.0.0

# 机器学习 (AI预测功能，可选)
scikit-learn>=1.0.0

# 高效数据存储 (可选)
pyarrow>=6.0.0
openpyxl>=3.0.0

# WiFi性能测试
speedtest-cli>=2.1.3

# 如果需要更高级的网络诊断功能（可选）
# scapy>=2.5.0  # 高级网络包分析，需要时取消注释
# dnspython>=2.3.0  # 高级DNS查询，需要时取消注释

# EXE打包工具（开发环境）
# pyinstaller>=5.13.0  # 将Python程序打包为exe，打包时需要安装